Refactor audit topic to topic package
