+++ Core/src/main/java/com/arivanamin/healthcare/backend/base/domain/topics/AuditTopics.java
@@ -1,4 +1,4 @@
-package com.arivanamin.healthcare.backend.base.domain.audit;
+package com.arivanamin.healthcare.backend.base.domain.topics;
 
 public interface AuditTopics {
